@@ -1,6 +1,6 @@
 {
   "name": "simple-update-notifier",
-  "version": "1.0.7",
+  "version": "1.1.0",
   "description": "Simple update notifier to check for npm updates for cli applications",
   "main": "build/index.js",
   "types": "build/index.d.ts",
@@ -15,7 +15,7 @@
     "node": ">=8.10.0"
   },
   "scripts": {
-    "test": "jest src",
+    "test": "jest src --noStackTrace",
     "build": "rollup -c rollup.config.js",
     "prettier:check": "prettier --check src/**/*.ts",
     "prettier": "prettier --write src/**/*.ts",
@@ -28,22 +28,22 @@
     "semver": "~7.0.0"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.18.2",
+    "@babel/preset-env": "^7.19.1",
     "@babel/preset-typescript": "^7.17.12",
-    "@release-it/conventional-changelog": "^5.0.0",
-    "@types/jest": "^28.1.3",
-    "@types/node": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.29.0",
-    "@typescript-eslint/parser": "^5.29.0",
-    "eslint": "^8.18.0",
+    "@release-it/conventional-changelog": "^5.1.0",
+    "@types/jest": "^29.0.3",
+    "@types/node": "^18.7.18",
+    "@typescript-eslint/eslint-plugin": "^5.37.0",
+    "@typescript-eslint/parser": "^5.37.0",
+    "eslint": "^8.23.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^28.1.1",
+    "jest": "^29.0.3",
     "prettier": "^2.7.1",
-    "release-it": "^15.1.1",
-    "rollup": "^2.75.7",
+    "release-it": "^15.4.2",
+    "rollup": "^2.79.0",
     "rollup-plugin-ts": "^3.0.2",
-    "typescript": "^4.7.4"
+    "typescript": "^4.8.3"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
